Evaporation (mol / (m^2 s))

Evaporation (mol / (m^2 s))

E(T_leaf, pars, unitless)

Arguments

T_leaf
Leaf temperature in Kelvin

pars
Concatenated parameters (leaf_par, enviro_par, and constants)

unitless
Logical. Should function use parameters with units? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

Value

unitless = TRUE: A value in units of mol / (m ^ 2 / s) number of class numeric unitless = FALSE: A value in units of mol / (m ^ 2 / s) of class units

Details

The leaf evaporation rate is the product of the total conductance to water vapour (m / s) and the water vapour gradient (mol / m^3):

$$E = g_\mathrm{tw} D_\mathrm{wv}$$

If unitless = TRUE, T_leaf is assumed in degrees K without checking.

Examples

library(tealeaves)

cs <- make_constants()
ep <- make_enviropar()
lp <- make_leafpar()

T_leaf <- set_units(298.15, K)

E(T_leaf, c(cs, ep, lp), FALSE)

#> 0.005658263 [mol/m^2/s]