D_x: Calculate diffusion coefficient for a given temperature and pressure

D_x: Calculate diffusion coefficient for a given temperature and pressure

.get_Dx(D_0, Temp, eT, P, unitless)

Arguments

D_0

Diffusion coefficient at 273.15 K (0 °C) and 101.3246 kPa

Temp

Temperature in Kelvin

eT

Exponent for temperature dependence of diffusion

P

Atmospheric pressure in kPa

unitless

Logical. Should function use parameters with units? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

Value

Value in m\(^2\)/s of class units

Details

$$D = D_\mathrm{0} (T / 273.15) ^ {eT} (101.3246 / P)$$
According to Montieth & Unger (2013), eT is generally between 1.5 and 2. Their data in Appendix 3 indicate \(eT = 1.75\) is reasonable for environmental physics.

References

Monteith JL, Unsworth MH. 2013. Principles of Environmental Physics. 4th edition. Academic Press, London.

Examples

tealeaves:::.get_Dx(
  D_0 = set_units(2.12e-05, m^2/s),
  Temp = set_units(298.15, K),
  eT = set_units(1.75),
  P = set_units(101.3246, kPa),
  unitless = FALSE
)
#> 2.471125e-05 [m^2/s]