Calculate leaf energy balance

Calculate leaf energy balance

energy_balance(
  tleaf,
  leaf_par,
  enviro_par,
  constants,
  quiet = FALSE,
  components = FALSE,
  set_units = FALSE
)

Arguments

tleaf

Leaf temperature in Kelvin. If input is numeric, it will be automatically converted to units.

leaf_par

A list of leaf parameters. This can be generated using the make_leafpar function.

enviro_par

A list of environmental parameters. This can be generated using the make_enviropar function.

constants

A list of physical constants. This can be generated using the make_constants function.

quiet

Logical. Should a message appear about conversion from numeric to units? Useful for finding leaf temperature that balances heat transfer using uniroot.

components

Logical. Should leaf energy components be returned? Transpiration (in mol / (m^2 s)) also returned.

set_units

Logical. Should units be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

Value

A numeric value in W / m^2. Optionally, a named list of energy balance components in W / m^2 and transpiration in mol / (m^2 s).

Examples

library(tealeaves)

cs <- make_constants()
ep <- make_enviropar()
lp <- make_leafpar()
ep$T_sky <- ep$T_sky(ep)

T_leaf <- set_units(298.15, K)

energy_balance(T_leaf, lp, ep, cs, FALSE, TRUE, TRUE)
#> $energy_balance
#> 245.6976 [W/m^2]
#> 
#> $components
#> $components$R_abs
#> 1363.813 [W/m^2]
#> 
#> $components$S_r
#> 869.2087 [W/m^2]
#> 
#> $components$H
#> 0 [W/m^2]
#> 
#> $components$L
#> 248.9066 [W/m^2]
#> 
#> $components$E
#> 0.005658263 [mol/m^2/s]
#> 
#>