Make lists of parameters of leaf, environmental, or constant parameters

Make lists of parameters of leaf, environmental, or constant parameters

make_leafpar

make_enviropar

make_constants

make_leafpar(replace = NULL)

make_enviropar(replace = NULL)

make_constants(replace = NULL)

Arguments

replace
A named list of parameters to replace defaults. If NULL, defaults will be used.

Value

make_leafpar: An object inheriting from class leaf_par
make_enviropar: An object inheriting from class enviro_par
make_constants: An object inheriting from class constants

Details

Leaf parameters:

Symbol			R

Description

Units			Default

\(d\)			leafsize

Leaf characteristic dimension	m	0.1

\(\alpha_\mathrm{s}\)			abs_s

absorbtivity of shortwave radiation (0.3 - 4	\(\mu\)	m)	none	0.50

\(\alpha_\mathrm{l}\)			abs_l

absorbtivity of longwave radiation (4 - 80	\(\mu\)	m)	none

0.97		\(g_\mathrm{sw}\)

g_sw		stomatal conductance to H2O	(	\(\mu\)

mol H2O) / (m	\(^2\)	s Pa)	5

\(g_\mathrm{uw}\)			g_uw

cuticular conductance to H2O	(	\(\mu\)	mol H2O) / (m	\(^2\)

s Pa)	0.1		\(\mathrm{logit}(sr)\)

logit_sr		stomatal ratio (logit transformed)	none

0 = logit(0.5)		Symbol

Environment parameters:

Symbol			R

Description

Units			Default

\(P\)			P

atmospheric pressure	kPa	101.3246

\(r\)			r

reflectance for shortwave irradiance (albedo)	none	0.2		\(\mathrm{RH}\)

RH		relative humidity

none	0.50		\(S_\mathrm{sw}\)

S_sw		incident short-wave (solar) radiation flux density	W / m

\(^2\)		1000		\(S_\mathrm{lw}\)

S_lw		incident long-wave radiation flux density

W / m	\(^2\)			calculated

\(T_\mathrm{air}\)

T_air		air temperature	K	298.15

\(u\)			wind

windspeed	m / s	2

Constants:

Symbol			R

Description

Units			Default

\(c_p\)			c_p

heat capacity of air	J / (g K)	1.01

\(D_{h,0}\)			D_h0

diffusion coefficient for heat in air at 0 °C	m	\(^2\)	/ s	19.0e-06

\(D_{m,0}\)			D_m0

diffusion coefficient for momentum in air at 0 °C	m	\(^2\)	/ s

13.3e-06		\(D_{w,0}\)

D_w0		diffusion coefficient for water vapour in air at 0 C	m	\(^2\)

/ s	21.2e-06		\(\epsilon\)

epsilon		ratio of water to air molar masses	none

0.622		\(eT\)

eT		exponent for temperature dependence of diffusion	none	1.75

\(G\)			G

gravitational acceleration	m / s	\(^2\)

9.8		\(Nu\)

Nu		Nusselt number	none

calculated			\(R\)

R		ideal gas constant	J / (mol K)

8.3144598		\(R_\mathrm{air}\)

R_air		specific gas constant for dry air	J / (kg K)	287.058

\(\sigma\)			s

Stephan-Boltzmann constant	W / (m	\(^2\)	K

\(^4\)	)	5.67e-08		\(Sh\)

Sh		Sherwood number

none		calculated

Examples

library(tealeaves)

# Use defaults
cs <- make_constants()
ep <- make_enviropar()
lp <- make_leafpar()

# Replace defaults

ep <- make_enviropar(
  replace = list(
    T_air = set_units(300, K)
  )
)

lp <- make_leafpar(
  replace = list(
    leafsize = set_units(c(0.1, 0.2), m)
  )
)