R_abs: total absorbed radiation (W / m^2)

R_abs: total absorbed radiation (W / m^2)

.get_Rabs(pars, unitless)

Arguments

pars
Concatenated parameters (leaf_par, enviro_par, and constants)

unitless
Logical. Should function use parameters with units? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

Value

Value in W / m\(^2\) of class units

Details

The following treatment follows Okajima et al. (2012):

$$R_\mathrm{abs} = \alpha_\mathrm{s} (1 + r) S_\mathrm{sw} + \alpha_\mathrm{l} \sigma (T_\mathrm{sky} ^ 4 + T_\mathrm{air} ^ 4)$$

The incidient longwave (aka thermal infrared) radiation is modeled from sky and air temperature \(\sigma (T_\mathrm{sky} ^ 4 + T_\mathrm{air} ^ 4)\) where \(T_\mathrm{sky}\) is function of the air temperature and incoming solar shortwave radiation:

$$T_\mathrm{sky} = T_\mathrm{air} - 20 S_\mathrm{sw} / 1000$$

Symbol	R	Description	Units	Default

\(\alpha_\mathrm{s}\)	abs_s	absorbtivity of shortwave radiation (0.3 - 4 \(\mu\)m)	none	0.80

\(\alpha_\mathrm{l}\)	abs_l	absorbtivity of longwave radiation (4 - 80 \(\mu\)m)	none	0.97

\(r\)	r	reflectance for shortwave irradiance (albedo)	none	0.2

\(\sigma\)	s	Stefan-Boltzmann constant	W / (m\(^2\) K\(^4\))	5.67e-08

\(S_\mathrm{sw}\)	S_sw	incident short-wave (solar) radiation flux density	W / m\(^2\)	1000

\(S_\mathrm{lw}\)	S_lw	incident long-wave radiation flux density	W / m\(^2\)	calculated

\(T_\mathrm{air}\)	T_air	air temperature	K	298.15

\(T_\mathrm{sky}\)	T_sky	sky temperature	K	calculated

References

Okajima Y, H Taneda, K Noguchi, I Terashima. 2012. Optimum leaf size predicted by a novel leaf energy balance model incorporating dependencies of photosynthesis on light and temperature. Ecological Research 27: 333-46.

Examples

library(tealeaves)

cs <- make_constants()
ep <- make_enviropar()
lp <- make_leafpar()

tealeaves:::.get_Rabs(c(cs, ep, lp), FALSE)

#> 1363.813 [W/m^2]