g_tw: total conductance to water vapour (m/s)

g_tw: total conductance to water vapour (m/s)

.get_gtw(T_leaf, pars, unitless)

Arguments

T_leaf
Leaf temperature in Kelvin

pars
Concatenated parameters (leaf_par, enviro_par, and constants)

unitless
Logical. Should function use parameters with units? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

Value

Value in m/s of class units

Details

Total conductance to water vapor: The total conductance to water vapor (\(g_\mathrm{tw}\)) is the sum of the parallel lower (abaxial) and upper (adaxial) conductances:

$$g_\mathrm{tw} = g_\mathrm{w,lower} + g_\mathrm{w,upper}$$

The conductance to water vapor on each surface is a function of parallel stomatal (\(g_\mathrm{sw}\)) and cuticular (\(g_\mathrm{uw}\)) conductances in series with the boundary layer conductance (\(g_\mathrm{bw}\)). The stomatal, cuticular, and boundary layer conductance on the lower surface are:

$$g_\mathrm{sw,lower} = g_\mathrm{sw} (1 - sr) R (T_\mathrm{leaf} + T_\mathrm{air}) / 2$$ $$g_\mathrm{uw,lower} = g_\mathrm{uw} / 2 R (T_\mathrm{leaf} + T_\mathrm{air}) / 2$$
See .get_gbw for details on calculating boundary layer conductance. The equations for the upper surface are:

$$g_\mathrm{sw,upper} = g_\mathrm{sw} sr R (T_\mathrm{leaf} + T_\mathrm{air}) / 2$$ $$g_\mathrm{uw,upper} = g_\mathrm{uw} / 2 R (T_\mathrm{leaf} + T_\mathrm{air}) / 2$$
Note that the stomatal and cuticular conductances are given in units of (\(\mu\)mol H2O) / (m\(^2\) s Pa) (see make_leafpar) and converted to m/s using the ideal gas law. The total leaf stomtal (\(g_\mathrm{sw}\)) and cuticular (\(g_\mathrm{uw}\)) conductances are partitioned across lower and upper surfaces. The stomatal conductance on each surface depends on stomatal ratio (sr); the cuticular conductance is assumed identical on both surfaces.

Symbol			R

Description

Units			Default

\(g_\mathrm{sw}\)			g_sw

stomatal conductance to H2O	(	\(\mu\)	mol H2O) / (m

\(^2\)	s Pa)	5		\(g_\mathrm{uw}\)

g_uw		cuticular conductance to H2O

(	\(\mu\)	mol H2O) / (m	\(^2\)	s Pa)

0.1		\(R\)

R		ideal gas constant	J / (mol K)	8.3144598

\(\mathrm{logit}(sr)\)			logit_sr

stomatal ratio (logit transformed)	none	0 = logit(0.5)

\(T_\mathrm{air}\)			T_air

air temperature	K	298.15		\(T_\mathrm{leaf}\)

T_leaf		leaf temperature

K	input		Symbol

Examples

# Total conductance to water vapor

## Hypostomatous leaf; default parameters
leaf_par <- make_leafpar(replace = list(logit_sr = set_units(-Inf)))
enviro_par <- make_enviropar()
constants <- make_constants()
pars <- c(leaf_par, enviro_par, constants)
T_leaf <- set_units(300, K)

## Fixing boundary layer conductance rather than calculating
gbw_lower <- set_units(0.1, m / s)
gbw_upper <- set_units(0.1, m / s)

# Lower surface ----
## Note that pars$logit_sr is logit-transformed! Use stats::plogis() to convert to proportion.
gsw_lower <- set_units(pars$g_sw * (set_units(1) - stats::plogis(pars$logit_sr)) * pars$R *
                         ((T_leaf + pars$T_air) / 2), "m / s")
guw_lower <- set_units(pars$g_uw * 0.5 * pars$R * ((T_leaf + pars$T_air) / 2), m / s)
gtw_lower <- 1 / (1 / (gsw_lower + guw_lower) + 1 / gbw_lower)

# Upper surface ----
gsw_upper <- set_units(pars$g_sw * stats::plogis(pars$logit_sr) * pars$R *
                         ((T_leaf + pars$T_air) / 2), m / s)
guw_upper <- set_units(pars$g_uw * 0.5 * pars$R * ((T_leaf + pars$T_air) / 2), m / s)
gtw_upper <- 1 / (1 / (gsw_upper + guw_upper) + 1 / gbw_upper)

## Lower and upper surface are in parallel
g_tw <- gtw_lower + gtw_upper