tleaves: find leaf temperatures for multiple parameter sets

tleaves: find leaf temperatures for multiple parameter sets

tleaf: find leaf temperatures for a single parameter set

tleaves(
  leaf_par,
  enviro_par,
  constants,
  progress = TRUE,
  quiet = FALSE,
  set_units = TRUE,
  parallel = FALSE
)

tleaf(leaf_par, enviro_par, constants, quiet = FALSE, set_units = TRUE)

Arguments

leaf_par
A list of leaf parameters. This can be generated using the make_leafpar function.

enviro_par
A list of environmental parameters. This can be generated using the make_enviropar function.

constants
A list of physical constants. This can be generated using the make_constants function.

progress
Logical. Should a progress bar be displayed?

quiet
Logical. Should messages be displayed?

set_units
Logical. Should units be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

parallel
Logical. Should parallel processing be used via future_map?

Value

tleaves:

A tibble with the following units columns

Input:

abs_l	Absorbtivity of longwave radiation (unitless)

abs_s	Absorbtivity of shortwave radiation (unitless)

g_sw	Stomatal conductance to H2O (\(\mu\)mol H2O / (m^2 s Pa))

g_uw	Cuticular conductance to H2O (\(\mu\)mol H2O / (m^2 s Pa))

leafsize Leaf characteristic dimension	(m)

logit_sr	Stomatal ratio (logit transformed; unitless)

P	Atmospheric pressure (kPa)

RH	Relative humidity (unitless)

S_lw	incident long-wave radiation flux density (W / m^2)

S_sw	incident short-wave (solar) radiation flux density (W / m^2)

T_air	Air temperature (K)

wind	Wind speed (m / s)

Output:

T_leaf	Equilibrium leaf tempearture (K)

value	Leaf energy balance (W / m^2) at tleaf

convergence	Convergence code (0 = converged)

R_abs	Total absorbed radiation (W / m^2; see .get_Rabs)

S_r	Thermal infrared radiation loss (W / m^2; see .get_Sr)

H	Sensible heat flux density (W / m^2; see .get_H)

L	Latent heat flux density (W / m^2; see .get_L)

E	Evapotranspiration (mol H2O/ (m^2 s))

tleaf:

A data.frame with the following numeric columns:

T_leaf	Equilibrium leaf temperature (K)

value	Leaf energy balance (W / m^2) at tleaf

convergence	Convergence code (0 = converged)

R_abs	Total absorbed radiation (W / m^2; see .get_Rabs)

S_r	Longwave re-radiation (W / m^2; see .get_Sr)

H	Sensible heat flux density (W / m^2; see .get_H)

L	Latent heat flux density (W / m^2; see .get_L)

E	Evapotranspiration (mol H2O/ (m^2 s))

Examples

# tleaf for single parameter set:

leaf_par <- make_leafpar()
enviro_par <- make_enviropar()
constants <- make_constants()
tleaf(leaf_par, enviro_par, constants)

#> 
#> Solving for T_leaf ...

#>  done

#>         T_leaf         value convergence            R_abs              S_r
#> 1 301.4181 [K] -2.122124e-08           0 1363.813 [W/m^2] 907.9499 [W/m^2]
#>                  H                L                      E          Ar       Gr
#> 1 107.3552 [W/m^2] 348.5078 [W/m^2] 0.00794791 [mol/m^2/s] 0.004827203 788182.4
#>         Re       g_bw
#> 1 12778.08 0.02972824

# tleaves for multiple parameter set:

enviro_par <- make_enviropar(
  replace = list(
    T_air = set_units(c(293.15, 298.15), K)
  )
)
tleaves(leaf_par, enviro_par, constants)

#> Solving for T_leaf from 2 parameter sets...

#> # A tibble: 2 x 25
#>      abs_l    abs_s     g_sw     g_uw leafsize logit_sr        P        r
#>        [1]      [1] [umol/m^ [umol/m^      [m]      [1]    [kPa]      [1]
#> 1     0.97      0.5        5      0.1      0.1        0 101.3246      0.2
#> 2     0.97      0.5        5      0.1      0.1        0 101.3246      0.2
#> # … with 17 more variables: RH [1], S_sw [W/m^2], T_air [K], wind [m/s],
#> #   T_sky [K], T_leaf [K], value <dbl>, convergence <dbl>, R_abs [W/m^2],
#> #   S_r [W/m^2], H [W/m^2], L [W/m^2], E [mol/m^2/s], Ar <dbl>, Gr <dbl>,
#> #   Re <dbl>, g_bw <dbl>